EDIT ui platform now using JwkForm, update readme
The platform's public key section now uses the JwkForm so we have a
unified JWK section with the tool admin.

Turns out that blowing away the database made Passport not work anymore,
we can't request API tokens anymore. Needed to run the Passport
migrations, so I've added that into the readme.