FIX seeded data not auto incrementing id correctly
Apparently, if we specify id in our seeded data, postgres doesn't
update the auto-increment id and will reuse the ids we seeded. This
causes issues where the first few platforms/tools we add error out due
to duplicate ids because they were conflicting with the seeded
platform/tools.

To workaround this, if we don't specify the id in the seeded data, the
auto-increment gets properly updated.

Also changed the error for if the target tool doesn't exist. Instead of
throwing a 404, it now throws an LTIException.