ADD ui show shim configuration information
To help admins add the shim to other LTI platforms and tools, show the
configuration settings they should use. This is being pulled from the
database according to the lti.own_platform_id and lti.own_tool_id
configurations. So if the initial data entered is incorrect, then the
config shown here isn't going to be correct either.