ADD artisan console command to seed shim LTI info
The shim behaves as both an LTI platform and tool. We need the seed the
shim's platform and tool info into the database for the CI/CD builds.
The database seeders are meant for development and not for deployment so
I didn't want to use them. Instead, I've added an artisan console
command that'll use the shim's LTI config to automatically generate
platform and tool information.

We're generating this information based on the app.url config. This is
given in the .env file, so make sure app.url is correct in that file.

Since the lti:seed command, the database seeder, and the router all
needs the same information on the LTI routes, it was easier to
centralize their configuration in the config/lti.php file.