EDIT Dockerfile for kubernetes deployment
Except for the database, this should result in a single container that
will run the entire project. Using the php apache image for this
purprose. An example docker-compose file is provided in the
deploy-docker/ directory as a development aid for how kubernetes should
deploy this. The docker-compose file needs to be copied to the parent
directory in order to work.

The deploy-env file must be edited in order to reflect the correct app
url and the postgres server information. Of particular note is the app
key, which should definitely be changed if deployed to prod.

The custom entrypoint sets up the database, generate oauth keys, seed
data, etc.

Note that the container expects these following environment variables to
be set:

* DB_HOST - Database hostname, entrypoint needs to wait for the db to spin
    up before running migrations.
* DB_PORT - Database port.
* ADMIN_NAME - Admin name, entrypoint will create an admin user if it
    doesn't exist already.
* ADMIN_EMAIL - Admin email, this field is checked for duplicates.
* ADMIN_PASSWORD - Admin password.

One strange issue I encountered is that the built image kept trying to
serve js and css from port 8080. I think it has to do with some cached
results from my dev laradock being passed into the build, as that seems
to correspond with the hot module reload feature of laramix. I have no
idea where this cached/compiled result is located so the only way to
solve it was to let `npm run production` run on the entire project and
copying the results into the final image.