ADD custom passport install, deploy-docker readme
The stock passport:install command does not check if you already have
oauth_clients entries before creating new ones. This results in extra
clients being created everytime the container spins up. The new
passport:custominstall command adds this check to prevents the creation
of extra OAuth clients .

Added a readme to explain the contents of the deploy-docker/ directory.

Turned debug off in deploy-env.

Made it so that docker-compose can be run from the deploy-docker/
directory. It needed to be copied to the root source directory
previously.