ADD gitlab CI/CD configuration, fix .env not work
Enabling manual kubernetes deployment for the master branch and tags. We
are not building/pushing the image using gitlab since Dockerhub's
integration with github is doing that for us. This is why the deployment
is manual, as I couldn't figure out how to make gitlab deploy after
Dockerhub's build is complete.

Since .env was compiled when the image was built, when we change it in
helm charts, the change was not picked up. We need to regenerate the
config in entrypoint in order to changes in .env to be picked up.