FIX gitlab ci pulling image from the wrong tag
CI_APPLICATION_TAG overwrites the tag coming from the charts to the
wrong values. Changed it to pull from latest same as the charts.

I just remembered that this was originally so we can separate production
from staging deployments. Since the shim will only have staging
deployments on kubernetes, we shouldn't need to worry about breaking the
ci for the prod build with this change.