FIX not generating https urls and shim info error
The route() helper used in blade templates was generating http instead
of https urls in our staging deployment. This is because Laravel
determines whether to generate http or https urls based on the request.
But in the kubernetes deployment, https is applied and stripped by the
ingress, so Laravel only always see http requests. Added a FORCE_HTTPS
config to .env that can force Laravel to generate https urls.

Also fixed a js console only error with the shim info panel. It
complains that variables aren't defined because the api call to retrieve
the shim config hasn't succeeded yet, so the info is empty. Simple
workaround to display a loading text until the api call has returned.