EDIT platform/tool table add iss & oauth_token_url
Need to add columns to support LTI service requests.

Add the 'iss' column to tools table. This is the OAuth issuer field and
is usually just the tool's URL.

This is now necessary because tools are the ones that sign and send out
JWTs when it comes to LTI service requests. In LTI launches, it is the
platform that identifies itself via iss and now it is the tool that
needs to identify itself via iss.

Add the 'oauth_token_url' field to the platforms table.

LTI service requests needs to be authorized by an OAuth access token.
This is the endpoint on the platform we should go to in order to request
an access token.

Also edited seeds to add these new fields. Note that the seeded values
comes from documentation but I haven't confirmed if they actually work
yet.

TODO: this leaves the admin UI outdated as we'll have to add these
new fields to the forms.