ADD oauth access tokens for LTI service auth
LTI service calls are authorized by OAuth2 access tokens. This commit
starts implementing the endpoint for requesting this access token. In
particular, LTI 1.3 requests access tokens using the client credentials
grant flow via JWT as specified in RFC 7523.

Unfortunately, while Laravel Passport's OAuth implementation supports
client credential grants, it does not support the JWT auth that we need.
So we can't just use Passport for this.

The LTI tool requests a token by signing a JWT that we can verify is
from the tool. This commit verifies that the request is valid but does
not yet actually issue a valid token.

Tokens are scoped according to what LTI service they want to access.
Requests for services that the shim does not support are rejected.