EDIT oauth access token now generates a JWE
Use an encrypted JWT (JWE) for access tokens. Created an AccessToken
class for generating this JWE.

Honestly, I'm not sure what kind of data we need to put into the access
token for verification. It just occurred to me that we need some way of
telling which tool is talking to which platform. I'm thinking this could
be done when we rewrite the service endpoints in the launch id_token,
but until that's actually implemented, I'm not sure if it'll work out.

So this implementation is sort of a rough outline that should let us
proceed to the next stage and refine as needed.

While I could've used a JWS instead of a JWE, there's a possibility that
we might leak some data in a JWS. Opting for a JWE means that's one less
thing to worry about.