FIX broken deployment due to migrations
Quick but lazy fix. Deployment was broken cause I deleted the
platform_clients table migration completely. So for now, just drop all
the tables before running the migrations. This is ok for now since we're
not in production yet, but once in prod, we'll have to take care that
migrations can be run on existing data. I didn't want to have to deal
with that until we have to.

Also needed to fix the LTI info seed to add a value for the platform's
new shim_client_id field.