EDIT AccessToken can now request tokens
AccessToken can now request access tokens from other platforms. This is
just a form POST request to the platform access token endpoint with a
JWT signed by the shim's tool side.

Laravel has a http client that's just a thin wrapper around the Guzzle
library. We needed to add the Guzzle library to composer in order to use
it.

Note that the JWT Token Identifier (jti) that we're using is just hard
coded, it needs to be replaced with a real implementation.