ADD relay Name & Role Service Provisioning request
Names & Role Provisioning Service (NRPS) allows LTI tools to request
user enrolment information from the platform. With the NrpsFilter in
place, we rewrote the NRPS endpoint to one on the shim. This commit
implements actual functionality.

When an NRPS request arrives on the shim's platform side, we pass it off
to the shim's NRPS tool side. The tool side look up the stored NRPS
info, and then request an access token from the target platform. It then
uses the access token to make the NRPS request. The returned NRPS data
is then passed as is back the shim's NRPS platform side, which passes it
back to the originating tool.

Note that a lot of things are missing right now:
* It is currently just relaying the data without filtering.
* Various nonces (e.g. jti) are not implemented.
* It can't handle paging parameters passed by GET requests.