EDIT PlatformOAuthToken/ToolLaunch, minor refactor
Moved PlatformOAuthToken out of its platform/ directory. I expected it
to be more complicated than it turned out to be, so no need for the
additional directory for organizational reasons.

Use JwsUtil in ToolLaunch. The JwsUtil was originally based on the code
in ToolLaunch anyways. Replaced it with JwsUtil to keep code duplication
down.

Also removed various CompactSerializer imports that's no longer needed.