MOVE lti/Filters to lti/Specs/Launch/Filters
The original idea of having generalized filters didn't pan out.

LTI launches come with an LTI session and I used that session in all the
filters. LTI service calls do not have a session. This means that I
couldn't reuse the same filters. So it's more appropriate to consider
all the existing filters to be launch filters.

This doesn't mean that filters can't share code though. We might want to
generalize the code for mapping between real and fake users or
something. So the top level lti/Filters directory could make a comeback.
I get the feeling that most code sharing would be done in the model
though.