ADD nrps course context filter
Create the first NRPS filter which also sets up the the basic filter
structure. It's similar to the launch filter in that the Platform side
does the filtering and there's also a FilterInterface.

Part of the data returned by NRPS is the course context. This filter
rewrites the course context so that it uses the fake course id.

Note that currently, only the course ID is kept, fields like the course
title are dropped.