ADD NrpsUrlFilter, rewrite NRPS endpoint to shim
In the NRPS response, the 'id' field is the original platform NRPS
endpoint that was called. We need to rewrite this so that it points to
the shim's endpoint.

I'm actually not sure if the id field is required to be an NRPS endpoint
url, it's not explicitly stated in the specs but is shown in the
examples. So just in case, I've built a backup that just replaces the
'id' with a random string if it's not an NRPS endpoint we recognize.