ADD MemberFilter for NRPS, anonymizes users
LTI launches only has 1 user. NRPS calls have a list of users. I wanted
to avoid slow downs caused by too many sql queries, so tried to use bulk
queries. But it seems that Eloquent has no bulk query functionalities,
at least none that I can find that doesn't involve relations, making the
implementation a bit more messy than I'd like.

Both LtiFakeUser and LtiRealUser has a function that, given a set of
users, returns the associated database entry for those users, creating
them where necessary. I couldn't find a function to do this in the docs,
so had to implement it. It would've been nice to make it a shared
function, but there was sufficient differences that I didn't see a way
to do it, so each model has their own implementation. The implementation
for LtiRealUser lives in the MemberFilter though.

Since the LTI launch's UserFilter has similar functions to the
MemberFilter, I did experiment with using an AbstractUserFilter class to
share code between them. However, the shareable code didn't end up being
a whole lot, as most of it ended up in the models instead, so I ditched
that approach.

It turns out that login_hint in the lti_real_users table was a lot less
important than I initially thought. Since login_hint is not used at all
in LTI service calls, we can't use it as an additional lookup key. On
the other hand, the `sub` field can be treated as the primary lookup
key. According to spec, in the NRPS results, the `user_id` field should
be the user's `sub` given in the launch.