EDIT NRPS allow pagination and filtering params
NRPS spec allows the use of GET params to manipulate the response:
* limit - number of members to return, optional support for platforms
* role - only members the matching role will be returned

The original NRPS URL might have additional GET parameters in addition
to the NRPS params, so we need to be careful to preserve those
parameters. Conversely, when we convert the NRPS URL to an URL on the
shim, we only want the NRPS params and can ditch the other params.

This means that we could really use a library that can easily manipulate
URLs. So I've added the League\Uri libraries to composer. The library
requires the intl php extension for dealing with unicode URLs and the
gmp extension for dealing with IPs. The Dockerfile was updated to
install the intl extension, the gmp extension was already present.

Since we can now add parameters to the NRPS URL, NrpsUrlFilter cannot
count on retriving the Nrps entry by the URL. This was solved by
changing the FilterInterface to pass in the Nrps model.

TODO: Pagination URLs are actually passed in the HTTP header, will need
to handle them somehow.