ADD NRPS pagination filter
NRPS pagination is done through the 'link' http header following
RFC5988. This requires a bit of special handling in filter to work.

ToolNrps pulls out the link header and return it as part of the NRPS
response. PlatformNrps then applies all the filters in the same way as
for other params, this includes the pagination filter which rewrites the
pagination URLs to links on the shim. PlatformNrps then pulls link out
of the filtered params and place it in the header of an Response object.
This lets us return a response with the correct headers without having
to have additional code on the controller side.

The pagination links are treated like any other NRPS endpoint and we
create an entry in the nrps table for it.

I couldn't find a library that writes link headers for us, so I've
rolled a basic implementation. It only supports writing the 'rel' link
param and the link itself, hopefully this will be sufficient for our use
cases.

I noticed that Canvas, by default, only returns 50 users at a time. So
even though NRPS paging is an 'optional' mechanism, it's required if we
want to use it with Canvas courses that have more than 50 users. I think
this doesn't strictly follow the NRPS spec, which says "by default all
the current mememberships will be returned". But having a default limit
makes more sense technically anyways, since some of our combined section
courses can have thousand+ users.

This commit marks NRPS as functionally complete. A few things were not
implemented:

* Not supporting "2.4.3 Membership differences" because Canvas doesn't
  support it.
* Not supporting "3 Resource Link Membership Server" because I have no
  idea what the spec's expected use case is. Also shouldn't matter since
  Canvas doesn't seem to support it.