ADD AccessTokenTest, tests requesting access token
Only tests the codepath for requesting access tokens from a platform.
The other functionalities of AccessToken should be tested in
OAuthTokenTest.

It tests that the access token request has the correct parameters and
that the JWT is valid.

This test is sort of an prototype for using some of Laravel's more
advanced features. We're faking HTTP responses so that we don't need to
bring up an actual LTI platform for testing. We're using PHPUnit's
setUp() method to make sure the database and fake HTTP response are
properly configured before every test method's invocation. Note that we
must call the parent setUp() or tearDown() methods or we'll get errors.
We're using the default database seed that I've only been using for
manual testing so far.