ADD NrpsTest, basic testing of NRPS
As a start, NrpsTest makes sure we can do a successful NRPS call. As I
don't want to have to spin up an LTI platform to run these tests, the
responses from the platform are faked. This includes the access token
request and NRPS request. The access token isn't a proper access token,
but is fine for now since we still haven't actually implemented actual
access token verification yet.

The test mainly makes sure that users are entered into the database and
that properly filtered users are returned.

Not yet tested: Passing through role and limit queries. Filtering the
pagination mechanism. Also, NRPS url passed in 'id' filter.

The database setup needs to be performed in setUp() so we can add these
additional tests.