EDIT NrpsTest now tests limit & role queries
Nrps calls have append 'limit' and 'role' queries to the URL for
pagination and filtering purposes.

Since we don't have an actual LTI platform, we don't have to test
whether the limit and role is applied. We can get away with just making
sure that the queries are properly passed through the filtering chain.

Part of this involved moving the database setup from the member
filtering test to the shared setUp() function that is called before
every test.