EDIT NrpsTest now tests pagination header filter
NRPS pagination URLs are passed in the "link" header. These urls need to
be rewritten to shim URLs to make sure that we can filter those requests
too.

The test checks that URLs are rewritten and that the corresponding NRPS
entries are in the database.

This is going to be a bit fragile as I'm using hardcoded values to test
against, but I didn't want to have to deal with link header parsing in
tests. Will change it if it turns out be too fragile.