ADD Nonce implementation, ToolLaunch now uses it
A simple nonce implementation. Storage is handled by Laravel's Cache
mechanism. I've configured a database backed 'nonce' store that uses the
new 'nonce' table. Nonces currently have a default validity of 1 hour.

The nonce is currently 16 bytes of randomness generated by php's
random_bytes(). This should be secure enough without adding too many
characters to our requests (a worry for GET request compatibility).

I didn't really find any good guidance for how long nonces should live
for, so if anyone knows, feel free to change it.

One problem with Laravel cache is that it doesn't automatically clear
expired items. The expiration check only happens when you try to get the
cached item, so if you never get the item again, as is often the case
with nonces, it'll just stay in the database.

The ToolLaunch was the first place where I encountered the need for a
nonce, so that's where I've implemented the first nonce enforcement.
I've also added relevant tests to test that nonces are properly
validated. In the process, I found that some requirement parameters in
the id_token wasn't being checked, so I've beefed that up.

Realized the Tool is the one that supplies the nonce, so the
PlatformLaunch only needs to pass it on and don't need to validate it.

Next is AccessToken's jti which is basically a nonce.