EDIT AccessToken, PlatformOAuthToken, jti nonces
The JWT Token Identifier (jti) is basically a nonce for JWT that
protects against replay attacks.

In AccessToken, when sending a request for access tokens, we need to
generate a jti but don't need to do enforcement (up to the platform if
they want track jti and enforce it).

In PlatformOAuthToken, we need to track the jti given to us and protect
against replay. This required us to modify Nonce so that it'll store
nonces not generated by us, but otherwise, enforcement is the same as
nonces we generated.

OAuthTokenTest was overhauled to allow for easy testing of jti replay
protection. I noticed that it also allowed us to split up the tests
more, so there was a lot of changes.

The AccessTokenTest just checks that we properly store the jti as a
nonce.

Updated readme to talk about how to run individual tests since running
them all to debug one test isn't bearable anymore.