ADD access token caching
Instead of sending an access token request every time we need to make an
LTI service call, it's better to cache the access tokens to save on some
http calls. Since the LTI spec recommends access tokens to have a
lifetime of 1 hour, we should expect that they can stick around for a
little bit.

The access token's expiration time is sent as part of the access token
response in the 'expires_in' param. This is the number of seconds that
the access token is valid for.

If the access token has a very short lifetime, then I'm guessing that
the platform doesn't want us to cache them, so I've put in a minimum
threshhold that has to be met for us to cache it. This is currently 60
seconds.

Cache keys can only be 255 characters. Since we want to be able to
retrieve access tokens by platform and scopes, and scopes are lengthy
uris, there's a good chance it won't fit in the 255 char limit. So I've
mapped each supported scope to an int and are using that int as part of
the cache key instead. This was made easier by moving the scope validity
check into AccessToken from where it was in PlatformOAuthToken
previously.

The AccessToken class is getting a bit crowded. It might be time to move
away from all static methods, but I haven't got a good idea for how to
refactor it yet.

Tests were also added for access token caching.