EDIT PlatformNrps now verifies access tokens
We now actually check to see if the access token used is valid.

NrpsTest wasn't sending the authorization header in any of the requests,
so this was fixed. Added tests to make sure that invalid access tokens
are rejected.

We should also make sure that the access token has permission to access
the scope they're in. However, since we only have 1 valid scope right
now, it's a pointless test. So we should add this in after the
assignment and grades spec is implemented.

Can't think of a good way to test expired access tokens, so haven't done
that.