FIX authorization header can't be case sensitive
Found out that Laravel cares about capitalization when getting the
bearer token. Specifically, the bearer string.

This works: `Authorization: Bearer`
This doesn't work: `Authorization: bearer`

According to spec, the bearer shouldn't be case sensitive. So now I'm
using a custom getter that isn't case sensitive.

Added a test to cover this fix.

Discovered this issue when I tried out the Ltijs example server as a
replacement for the lti1.3 php example tool.