FIX jwt with floating point timestamp not working
The library we're using for JWTs only support ints for timestamps.
According to the spec, floating point timestamps are actually allowed.
So we have to do our timestamp validation. This validation method is in
the JwsUtil class.

Since we're not relying on the correctness of the library anymore, we
now need to test our timestamp validation.

OAuthTokenTest was updated with a more comprehsive test of timestamp
validity. But the corresponding test of the tool launch only tests for
expiration. Also, none of them test floating point timestamps. I didn't
want to reimplement JWT generation myself, so need to find some other
way to generate JWTs with floating point timestamps. Since the Ltijs
example server always generates floating point timestamps, I'm relying
on manual tests being able to pick this up for now.