EDIT lti launch views now all auto-submit
Each stage on in the LTI launch process was manually submitted for
easier debugging. Now that I'm pretty sure the launch works fine, I'm
converting them to auto submit.

I wrote the views back when bootstrap was still being included, so they
had bootstrap css classes. These are no longer necessary since the basic
layout we're using don't include any of those things.

Renamed the basic layout to autosubmitform, to make it more clear what
it's meant to do. All it does is look for the element with the id
autoSubmitForm and submit it with stock javascript. Hopefully, this
results in a very fast load time.