Revert "REMOVE platform_clients table, merge to platforms"
This reverts commit cbd17bc6f724a3d8aa864c76d3b0602002b58c4b.

Turns out we do need to have a client_id table. If each tool is to have
a separate entry on a platform, then each tool will need its own
client_id. Even though all of them point to the shim, we need to have
some way to determine what tool to go to, and the client_id is a better
way than using a custom parameter, imo.