EDIT ToolLaunch use PlatformClient get client_id
Long story short, we no longer need to use a custom parameter to tell
what the target tool we want to launch to is. There's two ways to tell
what target tool we want:

    1. OIDC optional parameter client_id. If given, we can easily
    retrieve the PlatformClient client_id and through that, get the
    tool.

    2. Unique target_link_uri. If client_id is not available, then
    hopefully we've configured a unique target_link_uri that tells us
    which tool we want to go to. This is currently just the regular shim
    target_link_uri with a query 'target_tool_id' that tells us the id
    of the tool we want.

TODO: the admin UI needs to be updated so guide admins through this
configuration.

The state used in the auth request used to contain a few parameters that
have been replaced by the PlatformClient id. It's really convenient that
PlatformClient ties both platform and tool together.

This was a pretty extensive change, so a lot of test cases were also
updated. Taking advantage of the database seeding to avoid doing a lot
of the manual setup I was doing before. Unfortunately, this does seem to
have slowed the tests down a bit.