EDIT LtiFakeUser add foreign key to CourseContext
We want each course to have different fake users. As it is currently,
only each tool has different fake users. Adding this foreign key to
CourseContext should let us let each real user have a different fake
user in each course on each tool.

This change triggered a bunch of changes as a result:

This means that we had to move the CourseContext migration up so
that it gets created before the LTI user tables.

We also had to add CourseContext to LtiSession as an easy way to provide
the CourseContext to fake users. Unfortunately, the code for this area
isn't done as well as I'd like. We now create the CourseContext in
ToolLaunch, instead of later on in the PlatformLaunch filters. This
results in calling the CourseContextFilter in the ToolLaunch, which
really doesn't feel right, as PlatformLaunch is supposed to be the one
that filters all the data. Can't think of a better way to do it though,
so leaving it alone for now.