EDIT launch midway now has user lookup
Instructors sometimes need to figure out who the real user behind a fake
user is, this is provided by the midway page. There's a countdown on the
midway page where if the instructor does nothing, it'll just continue on
to the tool as usual. If the instructor cancels the countdown, then they
activate the the lookup tool.

Using vue-good-table to display the mapping. This does mean that the
entire class roster is sent in the response in the background. This
might bog down in large classes and technical users would be able to
grab this list.

TODO:
* The look up tool currently receives too much information about the
users. We need to filter the information sent to a minimum.
* Email needs to be replaced by student number.
* Only instructors and administrators should be able to view this tool,
currently it is available to everyone.