EDIT MidwayLaunch lookup tool restricted by roles
Only institution admins and course instructors can lookup users.

RoleVocabulary is a helper intended to deal with role URIs used by the
LTI spec. It takes care of building the role URIs and checking that the
user has the right role for accessing the lookup tool. It has an
associated unit test.

The midway page is now split into two types. The original auto submit
page and the lookup tool page. Instructors and admins will pause on the
lookup tool page. Everyone else will see the auto submit page that sends
them directly on to the tool.