EDIT midway lookup send less user data to vue
The data passed to the vue components was basically the raw model query
results. To be safe, we're now filtering out most of the fields so that
only the fields that we might use are sent.

Ideally, this should be a restriction that lets us reduce the data we
actually query the database for. But I didn't see any quick way to do
this with Laravel, so we're using the Laravel Collection methods that
filter the data after the database query. This definitely isn't
scalable, but should suffice for now.