MOVE lti routes out of web into its own routegroup
LTI routes are now no longer in the web routes, but in its own lti
routes group. This is because when I found out that one of the
sideffects of disabling cookies broke the shim admin login. If I move
the lti routes into its own route group, I can re-enable cookies and fix
the admin page.

Also took the opportunity to no longer use config for the lti routes. I
initially started using them in order to share routes with the database
seeders. That turns out to be unnecessary because we can use the route()
helper in the seeders. So I've removed the route paths from
config/lti.php and the paths are now just directly in the routes/lti.php