ADD admin platform client list/add/edit/delete
Part of the changes on the backend is that platform clients are now
associated with a tool in additon to the platform. This requires a new
way to deal with them as the old interface assumes platform clients
belong to platforms.

I've split platform clients out into its own section, similar to
tool/platforms. The api endpoint is now "/api/platform-client" because
it isn't really properly a platform only thing. I'm not sure if this is
the best way to do it, honestly, but it is the easiest. The controller
has been fully fleshed out as compared to the barebones skeleton before.

Along with the new api, platform client also now has a corresponding
platformClient store.

The PlatformClientAdmin component doesn't split up into smaller
components, unlike ToolAdmin and PlatformAdmin. As I was writing it, it
seemed simple enough that it could remain a single component for now.