EDIT admin config info always show static info
The shim config info section requires selecting a tool for some
configuration values. This isn't necessary for the other fields which
are always the same, but these static config values are hidden along
with the dynamic ones. So I've now changed it to always show the static
values, with the dynamic values requiring a tool selection.

Minor wording and text color refactor.