EDIT PlatformLaunch move LtiSession to property
LtiSession was being initialized on every stage function, it could just
be moved into the constructor as a property and be functionally
equivalent and save us a few lines. Won't have to pass it separate to
the applyFilters() call either.

Also moved the constructor call on the controller into the try/catch,
since exceptions can be thrown in the constructor now when loading
LtiSession.