ADD logging around LTI launch and all requests
Using Laravel's logging mechanism, added an 'lti' and 'request' logging
channel. Both are daily logs stored in the default /storage/logs
directory. Changed log preservation to 366 days.

All HTTP requests are logged using the RequestLogger middleware to the
'request' logging channel.

The 'lti' logging channel can be accessed using the regular Laravel Log
interface or the new LtiLog helper. The LtiLog helper is just for
autoformatting the log output. Before that, I was creating a separate
methods in each class for it, so this hopefully saves a bit of work in
the long run.

Created an AbstractFilter class that instantiates an LtiLog instance
based on a class constant. So filters just need to extend it and change
the class constant to get an LtiLog instance to use.

The Midway codepath is a bit split, so a line of logging is still in the
MidwayController.

Most logging is debug logging. Info logging contains the start of a
launch and real/fake user mapping. Notice contains the last stage of a
launch, telling us what real/fake user is being launched to a course.

Still the Nrps logging to do.