ADD logging to NRPS
Similar to launch, NRPS logs initiation at the info level, and
successful completion at the notice level. In addition, the users synced
(fake and real) are logged at the info level, including their status and
roles.

Debug messages are the most common and are also generally similar to
launch, with pre-filter, filter, and post-filter messages being the
bulk of them. Nrps filters now also all inherit from AbstractFilter.

LtiLog now recognizes CourseContext, LtiRealUser, and LtiFakeUser. For
courseContext, it'll output the course's id, title, and label. For
users, it'll output the user's id and name. This ability means some of
the log messages can be simplified in the launch, so some of those
messages were modified.