EDIT midway lookup tool, make UI more friendly
UI improvements suggested by Letitia. Remove timer countdown to continue
on to the tool, as it is a stress inducing element that takes control
away from the user. Users will need an extra click to continue, but they
should get used to it. Also substituted the platform/tool names into
labels as appropriate in order to reduce confusion. New wording for
instructions and titles.

Also now allows reverse lookup from the platform side.

One thing to be aware of is that vue-good-table does not tolerate
missing properties well. In our case, some users don't have student
numbers (because they're not students), and vue-good-table complained.
So I had to add a computed property that processes the users list we got
to add empty string student numbers to users who don't have student
numbers.