FIX access token expiry time & unknown iss error
When an unknown iss is encountered, a php type error is thrown, this is
because the method's return type checking does not allow nulls. An
LtiException should be thrown instead, so I've changed the type to allow
nulls.

The access token expiry was set to a unix timestamp. This caused issues in
Webwork which took the timestamp to be seconds till expiry (e.g. 1 hour
would be 3600). This has now been fixed and the test was adjusted to
catch this in the future.