EDIT LtiLog now has a stream ID
The stream ID is mean to help us follow an LTI call from start to
finish. Previously, we were using the LtiSession id for that, but since
the LtiSession isn't generated until the last step of the ToolLaunch, we
miss out on the early parts of the LTI call.

The stream ID is generated at the start of the LTI launch and persisted
in the LtiSession when it is created.

Filters now need to be passed the LtiLog object so it can grab the
stream ID.

Midway departure has moved out of the MidwayController and into the
MidwayLaunch class to make passing the stream ID easier. I didn't set a
return class for getDepartureResponse() method because I couldn't figure
out what the global function redirect() returns. Using the Response
class like getArrivalResponse() didn't work.

NRPS calls are also given a stream ID.