ADD lti logging for access tokens
Realized that there were no lti logs around access token requests.

To enable the unified stream id for the lti logs resulted in passing the
LtiLog object around a lot. So there were a lot of changes just to make
sure the LtiExceptions are aware of what stream they're part of.

AccessToken needed to be converted to a non-static class in order to
handle the LtiLog properly.