EDIT launch & nrps, send fake user first/last name
Webwork operates on the first and last name structure, so it wasn't
happy when we only sent the "name" field". Since these are fake names
anyway, the shim can send first and last names just to make these apps
happy. This is sent in both the launch and nrps responses.

In LTI, first and last names are the given_name and family_name fields.

Note that the underlying database table hasn't changed. There's still
only the single "name" field. I'm just splitting on space to get the
first and last names. This is possible because Faker generated first and
last names are single words without spaces.