EDIT platform AuthRespTest to check NRPS claims
Overhauled the AuthRespTest to allow for multiple tests. This lets us
split the tests up into units. Was initially doing this to prepare for
AGS tests. But realized that NRPS claims weren't being checked in the
test, so adding those tests became the goal instead.

The test checks that a proper NRPS entry was created and that the NRPS
url is actually ours.