EDIT launch WhitelistFilter split opt/req claims
Split the launch WhitelistFilter so that we have separate lists for
required and optional claims. This is for deduplication purposes.
PlatformLaunch had it's own list of optional claims that it checks to
add to the claims we send to the filters. For easier maintenance, it's
better to have that list in one place, in the WhitelistFilter. This
change came from me being confused when I added an entry to the
whitelist but the claim didn't show up due to PlatformLaunch's own list
of optional claims also needing to be changed.

Added a debug lti log output for whitelists to tell us what claims
they've removed. This would've been helpful for debugging when I was
trying to figure this out.