EDIT AgsFilter to filter lineitems url
Like NRPS, we now have an AGS table to track AGS endpoints.

Note the difference between lineitem and lineitems. The plural one is
required to be present, while the singular lineitem one is for resource
links that point to a single assignment.

The AGS model is also based on the NRPS model.

I noticed that the NRPS table doesn't index on course_context_id, I've
added it to the index in the AGS table but don't want to touch NRPS for
now.

Unlike NRPS, there are multiple scopes available for AGS, so we need a
scopes field to track what capabilities are available for this endpoint.