EDIT ags table, lineitem split into its own table
Instead of lineitems and lineitem all being in the same ags table, it
makes more sense to have a separate table storing individual lineitem
entries. For some reason, I thought the lineitems + lineitem message in
the launch claims was special and needs to be together. Now I realized
due to how lineitem is retrieved, it will always be related to a
lineitems entry. So it's ok to split lineitem off to its own table.

Just realized that the AgsController wasn't committed, so the router
as it is probably doesn't work. I don't want to commit the AgsController
yet though as it's connected to everything else too.

Added debug logging so we know if an access token was retrieved from
cache or from platform.